ggds-backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.0

# Database (PostgreSQL)
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Email
email-validator==2.1.1
resend==2.4.0

# File Upload (S3 - placeholder for future)
boto3==1.34.162
python-magic==0.4.27

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
